Prohibir acceso a contactos que no les pertenecen

+++ delete.php
@@ -21,6 +21,14 @@ if ($statement->rowCount() == 0){
   return;
 }
 
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo("HTTP 403 NO Autorizado");
+  return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
 
 header("Location: home.php");
